requirements.txt
# Core dependencies
chromadb==0.4.22
transformers==4.51.3
torch==2.7.0
llama-index-core==0.12.33.post1
llama-index-embeddings-huggingface==0.5.3
llama-index-llms-ollama==0.5.4

# Document processing
docling==2.32.0
docling-core==2.31.0
docling-parse==4.0.1

# Utilities
tqdm==4.67.1
numpy==1.26.4
pyyaml==6.0.2

# Jupyter support
jupyter==1.1.1
ipykernel==6.29.5
notebook==7.3.2

# Optional but recommended
accelerate==1.2.1
sentence-transformers==4.1.0
